+++ src/game/server/gamecontext.h
@@ -734,7 +734,7 @@ public:
 	bool IsSpawnArea(vec2 Pos);
 
 	// saved
-	bool SaveCharacter(int ClientID, int Flags = 0, int Hours = -1);
+	bool SaveCharacter(int ClientID, int Flags = 0, float Hours = -1);
 	int FindSavedPlayer(int ClientID);
 	bool CheckLoadPlayer(int ClientID);
 	const char *GetSavedIdentityHash(SSavedIdentity Info);
@@ -746,7 +746,7 @@ public:
 	void ExpireSavedIdentities();
 	void RemoveSavedIdentityFile(const char *pHash, const char *pName);
 	
-	void SaveDrop(int ClientID, int Hours, const char *pReason);
+	void SaveDrop(int ClientID, float Hours, const char *pReason);
 
 private:
